Updated namespace  "Resource\Api"  was renamed to "Model"
