+++ src/Model/Tweet/CreateResponse.php
@@ -5,50 +5,39 @@
  * of the MIT license.  See the LICENSE file for details.
  */
 
-namespace FAPI\Boilerplate\Resource\Api\Tweet;
+namespace FAPI\Boilerplate\Model\Tweet;
 
-use FAPI\Boilerplate\Resource\ApiResponse;
+use FAPI\Boilerplate\\ApiResponse;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class IndexResponse implements ApiResponse
+class CreateResponse implements ApiResponse
 {
     private $message;
 
-    private $tweets;
-
     /**
      * @param string $message
-     * @param array  $tweets
      */
-    private function __construct(string $message, array $tweets)
+    private function __construct(string $message)
     {
         $this->message = $message;
-        $this->tweets = $tweets;
     }
 
     /**
      * @param array $data
      *
-     * @return IndexResponse
+     * @return CreateResponse
      */
     public static function create(array $data)
     {
         $message = '';
-        $tweets = [];
-
-        if (isset($data['tweets'])) {
-            foreach ($data['tweets'] as $item) {
-                $tweets[] = Tweet::create($item);
-            }
-        }
 
         if (isset($data['message'])) {
             $message = $data['message'];
         }
 
-        return new self($message, $tweets);
+        return new self($message);
     }
 
     /**
@@ -57,13 +46,5 @@ class IndexResponse implements ApiResponse
     public function getMessage(): string
     {
         return $this->message;
-    }
-
-    /**
-     * @return Tweet[]
-     */
-    public function getTweets(): array
-    {
-        return $this->tweets;
     }
 }
