+++ src/Model/Tweet/IndexResponse.php
@@ -5,39 +5,50 @@
  * of the MIT license.  See the LICENSE file for details.
  */
 
-namespace FAPI\Boilerplate\Resource\Api\Tweet;
+namespace FAPI\Boilerplate\Model\Tweet;
 
-use FAPI\Boilerplate\Resource\ApiResponse;
+use FAPI\Boilerplate\\ApiResponse;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class UpdateResponse implements ApiResponse
+class IndexResponse implements ApiResponse
 {
     private $message;
 
+    private $tweets;
+
     /**
      * @param string $message
+     * @param array  $tweets
      */
-    private function __construct(string $message)
+    private function __construct(string $message, array $tweets)
     {
         $this->message = $message;
+        $this->tweets = $tweets;
     }
 
     /**
      * @param array $data
      *
-     * @return CreateResponse
+     * @return IndexResponse
      */
     public static function create(array $data)
     {
         $message = '';
+        $tweets = [];
+
+        if (isset($data['tweets'])) {
+            foreach ($data['tweets'] as $item) {
+                $tweets[] = Tweet::create($item);
+            }
+        }
 
         if (isset($data['message'])) {
             $message = $data['message'];
         }
 
-        return new self($message);
+        return new self($message, $tweets);
     }
 
     /**
@@ -46,5 +57,13 @@ class UpdateResponse implements ApiResponse
     public function getMessage(): string
     {
         return $this->message;
+    }
+
+    /**
+     * @return Tweet[]
+     */
+    public function getTweets(): array
+    {
+        return $this->tweets;
     }
 }
